+++ crates/mysql/src/lib.rs
@@ -11,6 +11,6 @@ pub mod type_mapping;
 pub use connection::MySqlConnection;
 pub use di_extension::DbContextOptionsBuilderExt;
 pub use introspection::{introspect_mysql, DbColumn, DbTable};
-pub use provider::MySqlProvider;
+pub use provider::{MySqlProvider, MySqlTlsMode};
 pub use sql_generator::MySqlSqlGenerator;
 pub use type_mapping::MySqlTypeMapping;